Simple http static server
This is the first commit of this project.
I want to create a simple,light framework that give me the basically functionalities of MVC framework in less the 1000 lines of code.
In this first commit I created the htpp-server that render a index.html file.
There isn't route, there isn't controller or other things now.
For controller I mean 'handler' .
In this version of the framework there isn't static file, if front-end's developer has modified the html of index.html , if you refresh the browser you can watch the new features .
So this is the beginning of my project.
I need contributors, I want contributors .
So create pull-request or Issue or everything you want every time .