Serving Static File
Now the framework can server static file.
The url is '/static/{path_file}' and path_file mean the path of the file.
The Content Header of the response is done with MIME Type from WikiPedia
but it's enabled only for pdf,json,css,js,png,jpg.
Now I'm using readFileSync for read file and serving it for http-server.